Fixing scrollbar and fullscreen with SDL3 (kind of).
git-svn-id: https://svn.code.sf.net/p/previous/code/branches/branch_filesharing@1461 9ba66e4b-2233-44b8-a46a-7e93b2105c9e

+++ src/gui-sdl/dlgFileSelect.c
@@ -130,7 +130,7 @@ static int ypos = -1;				/* First entry number to be displayed. If -1, file sele
 						/* else we continue from the previous position when SDLGui_FileSelect is called again */
 static bool refreshentries;			/* Do we have to update the file names in the dialog? */
 static int entries;				/* How many files are in the actual directory? */
-static int oldMouseY = 0;			/* Keep the latest Y mouse position for scrollbar move computing */
+static float oldMouseY = 0;			/* Keep the latest Y mouse position for scrollbar move computing */
 static int mouseClicked = 0;			/* used to know if mouse if down for the first time or not */
 static int mouseIsOut = 0;			/* used to keep info that mouse if above or under the scrollbar when mousebutton is down */
 static float scrollbar_Ypos = 0.0;		/* scrollbar height */
